refactor: extracted ProfileCard logic into useProfileForm hook
